# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TESIS - Sistema de Clasificación de Riesgo",
        "description": "Documentación de la API del sistema de tesis",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/actividades-economicas": {
            "get": {
                "tags": [
                    "Actividad Económica"
                ],
                "summary": "Listar todas las actividades económicas",
                "description": "Obtiene todas las actividades económicas, incluyendo las eliminadas (soft delete)",
                "operationId": "4e9a361d181ddf551b00916b8b3fa83c",
                "responses": {
                    "200": {
                        "description": "Listado de actividades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Comercio minorista"
                                            },
                                            "funcion_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "estado": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Actividad Económica"
                ],
                "summary": "Crear una nueva actividad económica",
                "description": "Crear nueva actividad",
                "operationId": "c2fdcc57bb9b49a180284345199656d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descripcion",
                                    "funcion_id"
                                ],
                                "properties": {
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Servicios profesionales"
                                    },
                                    "funcion_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actividad creada correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
        "/api/actividades-economicas/{id}": {
            "get": {
                "tags": [
                    "Actividad Económica"
                ],
                "summary": "Obtener una actividad económica",
                "description": "Mostrar actividad específica",
                "operationId": "08e42d8f09c509344ec60c63e050be97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad encontrada"
                    },
                    "404": {
                        "description": "Actividad no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Actividad Económica"
                ],
                "summary": "Actualizar una actividad económica",
                "description": "Actualizar actividad",
                "operationId": "b6f987a1840ade57c5f0ddab93b3dc65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descripcion",
                                    "funcion_id"
                                ],
                                "properties": {
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Actividad actualizada"
                                    },
                                    "funcion_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actividad actualizada correctamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actividad Económica"
                ],
                "summary": "Eliminar (soft delete) una actividad económica",
                "description": "Soft delete (marca estado = 0)",
                "operationId": "6672fffea364378c1b83ba9fd9968acf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad eliminada"
                    }
                }
            }
        },
        "/api/actividades-economicas/{id}/restore": {
            "post": {
                "tags": [
                    "Actividad Económica"
                ],
                "summary": "Restaurar una actividad económica eliminada",
                "description": "Restaurar actividad eliminada",
                "operationId": "0ab74da4523ee58476f5cdeffe497b67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad restaurada correctamente"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingresa tu token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}